refactor(www): group sidebar menus
