+++ apps/www/components/sidebar/menu/utility.tsx
@@ -14,7 +14,7 @@ import { useRouter } from "@repo/internationalization/src/navigation";
 import { COMPANY_SOCIAL_PROFILES } from "@repo/seo/company-profiles";
 import { useTranslations } from "next-intl";
 import type * as React from "react";
-import { SidebarPreferenceSubmenus } from "@/components/sidebar/preference-menu";
+import { SidebarPreferenceSubmenus } from "@/components/sidebar/menu/preference";
 
 /**
  * Renders the sidebar utilities that belong inside the user dropdown in both
